# Complete Redis Streams Consumer Groups Implementation with Performance Validation