--- a/Cargo.toml
+++ b/Cargo.toml
@@ -19,6 +19,7 @@ lazy_static = "1.4"
 rand = "0.8.5"
 thiserror = "1.0"
 futures = "0.3"
+crossbeam-epoch = "0.9"
 tokio = { version = "1.0", features = ["rt", "time"] }
 uuid = { version = "1.3.0", features = ["v4"] }
 crossbeam = "0.8"