… docs/ROADMAP.md
… docs/STREAMS_SPECIFICATION.md
… docs/VALIDATION.md
… src/storage/commands/consumer_groups.rs
… src/storage/consumer_groups.rs
… tests/features/streams/test_consumer_groups_comprehensive.py
… tests/validate_consumer_groups.sh